refactor: Moved CFE callback into a separate hook
